secure_data_encryption/requirements.txt
streamlit==1.32.0
cryptography==42.0.5
passlib==1.7.4
python-jose==3.3.0 